requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.51.0
importlib_resources==6.4.0
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyFFTW==0.13.1
pyparsing==3.1.2
pyrtlsdr==0.3.0
pyrtlsdrlib==0.0.3
python-dateutil==2.9.0.post0
scipy==1.13.0
SimpleSoapy==1.5.1
SimpleSpectral==1.0.0
six==1.16.0
soapy_power==1.6.1
zipp==3.18.2
